requirements.txt
pymatgen>=2022.0.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
tabulate>=0.8.9
requests>=2.25.0
numpy>=1.20.0
scipy>=1.7.0
plotly>=5.3.0  
python-dotenv>=0.19.0  